backend/requirements.txt
fastapi
uvicorn
anthropic
langchain
langchain-anthropic
langchain-pinecone
pinecone-client
finnhub-python
supabase
python-dotenv
pydantic